requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
basemap==1.3.8
basemap-data==1.3.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
contourpy==1.1.1
cycler==0.11.0
fonttools==4.42.1
frozenlist==1.4.0
idna==3.4
ipinfo==4.4.3
kiwisolver==1.4.5
matplotlib==3.7.3
multidict==6.0.4
numpy==1.25.2
packaging==23.1
Pillow==10.0.1
pyparsing==3.1.1
pyproj==3.6.1
pyshp==2.3.1
python-dateutil==2.8.2
requests==2.31.0
scapy==2.5.0
six==1.16.0
urllib3==1.26.6
yarl==1.9.2
python-dotenv